+++ packages/server/api/test/unit/ai/ai-config.service.test.ts
@@ -27,13 +27,13 @@ jest.mock('../../../src/app/helper/encryption', () => ({
   },
 }));
 
-import { SaveAiConfigRequest } from '@openops/shared';
+import { AiProviderEnum, SaveAiConfigRequest } from '@openops/shared';
 import { AiApiKeyRedactionMessage } from '../../../src/app/ai/config/ai-config.entity';
 import { aiConfigService } from '../../../src/app/ai/config/ai-config.service';
 
 describe('aiConfigService.upsert', () => {
   const baseRequest: SaveAiConfigRequest = {
-    provider: 'OpenAI',
+    provider: AiProviderEnum.OPENAI,
     apiKey: 'test-key',
     model: 'gpt-4',
     modelSettings: { temperature: 0.7 },
@@ -84,24 +84,6 @@ describe('aiConfigService.upsert', () => {
       id: 'mocked-id',
     });
     expect(encryptStringMock).toHaveBeenCalledWith(baseRequest.apiKey);
-  });
-
-  test('should throw an error if provider is invalid', async () => {
-    const invalidRequest = {
-      ...baseRequest,
-      provider: 'InvalidProvider',
-    };
-
-    await expect(
-      aiConfigService.upsert({
-        projectId,
-        request: invalidRequest,
-      }),
-    ).rejects.toThrow('Invalid AI provider: InvalidProvider');
-
-    expect(findOneByMock).not.toHaveBeenCalled();
-    expect(upsertMock).not.toHaveBeenCalled();
-    expect(encryptStringMock).not.toHaveBeenCalled();
   });
 
   test('should update existing ai config if it exists', async () => {
