+++ packages/shared/src/lib/ai/config/index.ts
@@ -1,10 +1,11 @@
 import { Static, Type } from '@sinclair/typebox';
 import { BaseModelSchema } from '../../common/base-model';
+import { AiProviderEnum } from '../providers/index';
 
 export const AiConfig = Type.Object({
   ...BaseModelSchema,
   projectId: Type.String(),
-  provider: Type.String(),
+  provider: Type.Enum(AiProviderEnum),
   model: Type.String(),
   apiKey: Type.String(),
   providerSettings: Type.Optional(Type.Record(Type.String(), Type.Unknown())),
@@ -16,7 +17,7 @@ export type AiConfig = Static<typeof AiConfig>;
 
 export const SaveAiConfigRequest = Type.Object({
   id: Type.Optional(Type.String()),
-  provider: Type.String(),
+  provider: Type.Enum(AiProviderEnum),
   model: Type.String(),
   apiKey: Type.String(),
   providerSettings: Type.Optional(Type.Record(Type.String(), Type.Unknown())),
